Fix tracer disable/reenable code path

+++ source/layers/tracing/tracing_imp.h
@@ -122,7 +122,7 @@ struct APITracerContextImp : APITracerContext {
 
     ze_bool_t testForTracerArrayReferences(tracer_array_t *tracerArray);
     size_t testAndFreeRetiredTracers();
-    int updateTracerArrays();
+    size_t updateTracerArrays();
 
     std::list<ThreadPrivateTracerData *> threadTracerDataList;
     std::mutex threadTracerDataListMutex;
